requirements.txt
infi.systray
clipboard